refactor(build-cache): modularize cache functions and improve file handling
